feat(SVG): ReactJS SVG not displaying on mobile devices
SVG file is not showing up on mobile devices.

Added `xmlns:xlink="http://www.w3.org/1999/xlink"` in an attempt to fix the issue.

https://stackoverflow.com/a/27249490

```
The other problem is that Safari isn't particularly brilliant at rendering SVGs. However, it tends to do better when you embed them with an <iframe> or <object> tag instead of using <img>
```